Add setitem methods for NdArray and Variable
