+++ python/src/nnabla/_indexing.pxd
@@ -1,1 +1,2 @@
 cdef object getitem(object x, object key)
+cdef object setitem(object x, object key, object value)
