+++ python/src/nnabla/_nd_array.pyx
@@ -443,3 +443,6 @@ cdef class NdArray:
 
     def __getitem__(self, key):
         return IDX.getitem(self, key)
+
+    def __setitem__(self, key, value):
+        IDX.setitem(self, key, value)
